refactor: failing tests fix and improvements
